updates to allow csv forcing
