+++ test_data/Gage_01123000_csv/cat-11466.csv
@@ -1,0 +1,19 @@
+Time,U2D,V2D,LWDOWN,RAINRATE,T2D,Q2D,PSFC,SWDOWN
+2025-02-20 15:00:00,0.792,-2.814,191.328,0,266.9,0.001349,99149.5,539.232
+2025-02-20 16:00:00,0.109,-2.05,198.545,0,268.02,0.001483,98998.1,611.038
+2025-02-20 17:00:00,1.044,-2.056,202.387,0,269.13,0.001609,99029.5,650.483
+2025-02-20 18:00:00,0.702,-1.589,214.491,0,269.73,0.001823,98934.7,582.878
+2025-02-20 19:00:00,0.712,-2.757,239.591,0,270.03,0.002107,98758.1,441.591
+2025-02-20 20:00:00,0.699,-2.393,285.533,0,269.85,0.002417,98698.1,146.196
+2025-02-20 21:00:00,0.871,-2.21,285.905,1.99466558825634e-08,269.55,0.002408,98708.1,65.686
+2025-02-20 22:00:00,1.16,-1.65,285.135,1.56998922307139e-07,269.27,0.002408,98798.1,8.292
+2025-02-20 23:00:00,1.35,-1.556,282.701,1.37052268200932e-07,268.92,0.002354,98882.9,0
+2025-02-21 00:00:00,1.126,-1.475,281.208,4.10320035371114e-06,268.61,0.002307,98907.4,0
+2025-02-21 01:00:00,2.177,-1.808,281.035,8.98366215551505e-06,268.28,0.002286,98864,0
+2025-02-21 02:00:00,3.328,-2.844,280.598,7.67198162066052e-06,267.86,0.002214,98860.1,0
+2025-02-21 03:00:00,3.551,-2.953,274.01,3.49171727975772e-06,267.19,0.002025,98857.4,0
+2025-02-21 04:00:00,4.323,-3.511,237.324,1.60811384830595e-06,266.06,0.001849,98765.3,0
+2025-02-21 05:00:00,4.294,-3.625,222.341,5.74366549699334e-06,265.51,0.001769,98745.3,0
+2025-02-21 06:00:00,4.374,-4.081,260.361,1.01648311101599e-06,265.86,0.001773,98785.3,0
+2025-02-21 07:00:00,4.366,-4.145,266.781,3.52459181840459e-07,266.1,0.001782,98947.4,0
+2025-02-21 08:00:00,4.428,-2.926,247.543,0,265.87,0.001763,99026.7,0
