+++ test_data/Gage_01123000_csv/cat-11467.csv
@@ -1,0 +1,19 @@
+Time,U2D,V2D,LWDOWN,RAINRATE,T2D,Q2D,PSFC,SWDOWN
+2025-02-20 15:00:00,0.758,-2.851,192.956,0,267.26,0.00141,99640.4,537.832
+2025-02-20 16:00:00,0.523,-2.426,197.546,0,268.37,0.00151,99484.1,627.121
+2025-02-20 17:00:00,0.855,-2.042,206.172,0,269.57,0.001621,99527.7,641.23
+2025-02-20 18:00:00,0.596,-1.747,208.462,0,270.1,0.001835,99427.7,616.255
+2025-02-20 19:00:00,0.784,-3.035,236.852,0,270.3,0.002082,99247.7,430.428
+2025-02-20 20:00:00,0.618,-2.488,284.795,0,270.25,0.0023,99197.7,172.695
+2025-02-20 21:00:00,0.26,-2.371,287.52,0,269.9,0.002393,99207.7,45.438
+2025-02-20 22:00:00,0.62,-2.113,287.299,1.17783451969444e-06,269.76,0.002391,99297.7,7.782
+2025-02-20 23:00:00,0.4,-1.709,285.371,3.56382588506676e-06,269.62,0.002412,99377.7,0
+2025-02-21 00:00:00,0.884,-1.32,282.467,1.2050659279339e-05,269.39,0.0024,99407.7,0
+2025-02-21 01:00:00,1.973,-1.658,282.189,9.8840391729027e-06,269.03,0.002232,99368.2,0
+2025-02-21 02:00:00,2.919,-2.037,282.106,9.38357879931573e-06,268.63,0.002231,99358.2,0
+2025-02-21 03:00:00,3.371,-2.767,278.065,4.51473715656903e-06,268.06,0.002022,99348.8,0
+2025-02-21 04:00:00,4.306,-3.252,239.9,8.95100868092413e-07,267.06,0.001841,99249.3,0
+2025-02-21 05:00:00,4.184,-3.59,240.403,3.11436951960786e-06,266.54,0.001761,99234.1,0
+2025-02-21 06:00:00,4.032,-4.083,264.086,1.25582755572395e-06,266.73,0.001722,99274.1,0
+2025-02-21 07:00:00,3.918,-4.279,266.154,1.76601261614451e-07,267.02,0.001732,99444.1,0
+2025-02-21 08:00:00,4.363,-2.819,233.758,0,266.63,0.001706,99528.2,0
