+++ test_data/Gage_01123000_csv/cat-11468.csv
@@ -1,0 +1,19 @@
+Time,U2D,V2D,LWDOWN,RAINRATE,T2D,Q2D,PSFC,SWDOWN
+2025-02-20 15:00:00,1.049,-3.097,194.012,0,267.64,0.001388,99899.4,538.833
+2025-02-20 16:00:00,0.859,-2.472,197.599,0,268.77,0.00144,99761,629.704
+2025-02-20 17:00:00,1.042,-2.052,204.894,0,269.84,0.001545,99814.4,648.795
+2025-02-20 18:00:00,1.132,-1.756,206.135,0,270.41,0.001719,99699,620.776
+2025-02-20 19:00:00,1.363,-2.974,225.007,0,270.67,0.001942,99514.9,468.159
+2025-02-20 20:00:00,0.888,-2.582,274.586,0,270.66,0.002167,99474.4,228.357
+2025-02-20 21:00:00,0.415,-2.254,287.74,0,270.29,0.002314,99478.2,41.091
+2025-02-20 22:00:00,0.891,-2.019,287.287,9.33754620291438e-08,270.11,0.002297,99574.4,12.185
+2025-02-20 23:00:00,0.881,-1.731,285.052,1.20095103284257e-06,269.87,0.002256,99654.4,0
+2025-02-21 00:00:00,1.046,-1.468,282.259,4.83463645650772e-06,269.61,0.002251,99684.4,0
+2025-02-21 01:00:00,2.249,-1.431,282.579,7.62731997383526e-06,269.31,0.002213,99654.4,0
+2025-02-21 02:00:00,3.344,-1.753,282.767,1.77536730916472e-05,268.68,0.002173,99651.8,0
+2025-02-21 03:00:00,3.563,-3.079,278.598,4.73335740025504e-06,268.28,0.002007,99635.6,0
+2025-02-21 04:00:00,4.31,-3.274,252.871,1.34906053972372e-06,267.5,0.001835,99536.8,0
+2025-02-21 05:00:00,4.306,-3.639,243.142,2.44010925598559e-06,266.85,0.001743,99516.3,0
+2025-02-21 06:00:00,4.101,-4.076,255.891,1.66666666245874e-06,266.89,0.00169,99555.6,0
+2025-02-21 07:00:00,4.006,-4.283,250.621,1.26259971366949e-08,267.02,0.0017,99722.2,0
+2025-02-21 08:00:00,4.482,-3.241,242.697,0,266.98,0.00169,99816.3,0
