+++ test_data/Gage_01123000_csv/cat-11469.csv
@@ -1,0 +1,19 @@
+Time,U2D,V2D,LWDOWN,RAINRATE,T2D,Q2D,PSFC,SWDOWN
+2025-02-20 15:00:00,0.924,-3.005,193.576,0,267.4,0.001388,99653.3,538.293
+2025-02-20 16:00:00,0.731,-2.498,196.844,0,268.52,0.001454,99502.8,628.996
+2025-02-20 17:00:00,0.987,-2.078,205.213,0,269.64,0.001556,99558.1,645.443
+2025-02-20 18:00:00,0.904,-1.78,205.45,0,270.21,0.001756,99451.3,619.84
+2025-02-20 19:00:00,1.132,-3.03,229.556,0,270.44,0.001994,99265.5,447.186
+2025-02-20 20:00:00,0.778,-2.555,282.015,0,270.41,0.002216,99221.3,198.286
+2025-02-20 21:00:00,0.24,-2.379,287.328,0,270.03,0.002355,99231.3,39.195
+2025-02-20 22:00:00,0.823,-2.131,286.732,3.82554389943834e-07,269.87,0.00233,99321.3,11.062
+2025-02-20 23:00:00,0.793,-1.794,284.898,2.17054912354797e-06,269.68,0.002299,99401.3,0
+2025-02-21 00:00:00,1.009,-1.535,282.171,7.37064465283765e-06,269.45,0.002301,99431.3,0
+2025-02-21 01:00:00,2.077,-1.578,282.171,8.38033156469464e-06,269.1,0.002207,99398.6,0
+2025-02-21 02:00:00,3.143,-1.811,281.867,1.37969791467185e-05,268.57,0.002198,99389.7,0
+2025-02-21 03:00:00,3.503,-2.945,277.479,4.60539285995765e-06,268.09,0.002014,99379.7,0
+2025-02-21 04:00:00,4.284,-3.296,249.937,1.11024598936638e-06,267.2,0.001835,99281.2,0
+2025-02-21 05:00:00,4.267,-3.672,247.189,2.77587491837039e-06,266.68,0.001754,99259.7,0
+2025-02-21 06:00:00,4.092,-4.133,257.009,1.59250953402079e-06,266.72,0.001697,99299.7,0
+2025-02-21 07:00:00,4.038,-4.407,256.044,4.35377138785498e-08,266.93,0.001706,99463.9,0
+2025-02-21 08:00:00,4.514,-3.146,230.612,0,266.7,0.001693,99559.7,0
