+++ test_data/Gage_01123000_csv/cat-11470.csv
@@ -1,0 +1,19 @@
+Time,U2D,V2D,LWDOWN,RAINRATE,T2D,Q2D,PSFC,SWDOWN
+2025-02-20 15:00:00,0.707,-2.723,193.036,0,267.33,0.001411,99739.6,536.866
+2025-02-20 16:00:00,0.307,-2.24,204.368,0,268.41,0.001516,99579.6,566.782
+2025-02-20 17:00:00,0.921,-2.019,205.77,0,269.65,0.001621,99619.6,642.067
+2025-02-20 18:00:00,0.585,-1.678,210.385,0,270.2,0.001834,99519.6,611.884
+2025-02-20 19:00:00,0.701,-2.871,233.496,0,270.41,0.002103,99339.6,442.381
+2025-02-20 20:00:00,0.592,-2.391,286.201,0,270.33,0.002333,99288.7,146.062
+2025-02-20 21:00:00,0.404,-2.275,287.296,6.18914741679077e-09,270.02,0.002393,99298.9,50.32
+2025-02-20 22:00:00,0.79,-1.966,287.09,4.48279422471387e-07,269.89,0.00237,99388.7,7.308
+2025-02-20 23:00:00,0.685,-1.595,284.74,1.31408648940123e-06,269.66,0.00237,99469.6,0
+2025-02-21 00:00:00,0.95,-1.269,282.154,6.69180235490785e-06,269.43,0.002339,99498.7,0
+2025-02-21 01:00:00,2.047,-1.555,282.202,9.21615173865575e-06,269.05,0.002248,99458.7,0
+2025-02-21 02:00:00,3.09,-2.368,282.459,7.94950392446481e-06,268.62,0.002222,99448.7,0
+2025-02-21 03:00:00,3.464,-2.756,277.82,4.18467425333802e-06,268.04,0.002017,99446.3,0
+2025-02-21 04:00:00,4.33,-3.31,234.327,1.25920075788599e-06,267.02,0.001836,99346.3,0
+2025-02-21 05:00:00,4.282,-3.555,227.052,4.33934701504768e-06,266.44,0.001756,99328.4,0
+2025-02-21 06:00:00,4.205,-4.002,264.252,1.30372109197197e-06,266.7,0.001736,99368.7,0
+2025-02-21 07:00:00,4.099,-4.111,266.466,2.77777786550359e-07,266.93,0.001746,99538.7,0
+2025-02-21 08:00:00,4.323,-2.705,237.109,0,266.64,0.001713,99618.7,0
