+++ test_data/Gage_01123000_csv/cat-11475.csv
@@ -1,0 +1,19 @@
+Time,U2D,V2D,LWDOWN,RAINRATE,T2D,Q2D,PSFC,SWDOWN
+2025-02-20 15:00:00,0.727,-2.901,191.31,0,266.83,0.001361,99158.5,540.038
+2025-02-20 16:00:00,0.032,-2.052,201.812,0,267.91,0.001507,99008.2,593.963
+2025-02-20 17:00:00,0.796,-2.064,203.199,0,269.11,0.001651,99038.5,649.874
+2025-02-20 18:00:00,0.419,-1.59,214.166,0,269.64,0.001888,98938.6,576.776
+2025-02-20 19:00:00,0.432,-2.824,264.45,0,269.84,0.002132,98768.2,364.464
+2025-02-20 20:00:00,0.456,-2.505,285.814,0,269.74,0.002435,98708.2,142.887
+2025-02-20 21:00:00,0.791,-2.059,286.74,1.97764705944792e-07,269.43,0.002438,98718.2,64.705
+2025-02-20 22:00:00,0.647,-1.498,286.084,8.07674553016113e-07,269.22,0.002464,98808.2,7.133
+2025-02-20 23:00:00,0.542,-1.207,283.479,6.09909818649612e-07,268.98,0.002473,98898.1,0
+2025-02-21 00:00:00,0.931,-1.071,281.361,7.30867941456381e-06,268.63,0.002382,98911.1,0
+2025-02-21 01:00:00,2.015,-1.972,281.028,1.04325326901744e-05,268.23,0.002281,98861.5,0
+2025-02-21 02:00:00,3.169,-2.644,280.75,7.09363666828722e-06,267.92,0.002233,98863.6,0
+2025-02-21 03:00:00,3.523,-3.003,265.495,3.01202931041189e-06,267.1,0.002014,98861.1,0
+2025-02-21 04:00:00,4.443,-3.453,238.954,1.46059414873889e-06,265.97,0.001841,98763.7,0
+2025-02-21 05:00:00,4.335,-3.64,220.694,5.01193380841869e-06,265.46,0.001761,98743.7,0
+2025-02-21 06:00:00,4.353,-4.134,266.53,8.44531996335718e-07,265.89,0.001771,98783.7,0
+2025-02-21 07:00:00,4.28,-4.271,268.488,3.46592202049578e-07,266.13,0.001773,98951.1,0
+2025-02-21 08:00:00,4.441,-2.871,230.399,0,265.69,0.001755,99024,0
