+++ test_data/Gage_01123000_csv/cat-11476.csv
@@ -1,0 +1,19 @@
+Time,U2D,V2D,LWDOWN,RAINRATE,T2D,Q2D,PSFC,SWDOWN
+2025-02-20 15:00:00,0.665,-3.028,190.54,0,266.66,0.001391,99006.9,539.989
+2025-02-20 16:00:00,-0.152,-2.008,199.171,0,267.74,0.001547,98856.9,601.635
+2025-02-20 17:00:00,0.679,-2.01,201.387,0,268.94,0.00171,98886.9,654.517
+2025-02-20 18:00:00,0.311,-1.636,218.935,0,269.54,0.001956,98790.5,585.454
+2025-02-20 19:00:00,0.282,-2.702,272.56,0,269.74,0.002187,98620.5,333.321
+2025-02-20 20:00:00,0.379,-2.594,285.545,0,269.55,0.002433,98551.3,132.411
+2025-02-20 21:00:00,0.9,-2.006,286.027,1.41223821970016e-07,269.36,0.002422,98571.3,63.953
+2025-02-20 22:00:00,0.598,-1.122,285.98,7.68172071730078e-07,269.14,0.002485,98661.3,7.75
+2025-02-20 23:00:00,0.549,-0.949,283.509,7.61058743137255e-07,268.81,0.002472,98750.5,0
+2025-02-21 00:00:00,1.003,-1.141,281.179,7.73560896050185e-06,268.42,0.002333,98760.5,0
+2025-02-21 01:00:00,2.125,-2.375,280.881,9.92011882772204e-06,268.08,0.002262,98706.9,0
+2025-02-21 02:00:00,3.09,-2.571,280.187,8.32584919407964e-06,267.74,0.002211,98717.7,0
+2025-02-21 03:00:00,3.471,-3.085,270.524,2.11965652852086e-06,266.91,0.001996,98706.2,0
+2025-02-21 04:00:00,4.287,-3.264,237.939,1.68953647516901e-06,265.73,0.001823,98622,0
+2025-02-21 05:00:00,4.196,-3.469,252.282,5.35991739525343e-06,265.57,0.001764,98597.1,0
+2025-02-21 06:00:00,4.233,-4.135,265.965,8.16355111510347e-07,265.74,0.001763,98636.2,0
+2025-02-21 07:00:00,4.174,-4.205,269.347,4.37201549630117e-07,266,0.001769,98806.2,0
+2025-02-21 08:00:00,4.419,-2.897,236.034,0,265.58,0.00175,98876.2,0
